requirements.txt
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
idna==2.10
requests==2.24.0
selenium==3.141.0
urllib3==1.25.10
webdriver-manager==3.2.2
airtable-python-wrapper==0.15.1
twilio==2.24.0
